refector: Header 컴포넌트를 세부 컴포넌트로 분리 (#6)
